Fixed warnings
[git-p4: depot-paths = "//depot/C35/Students/ahadk/SD/Doomenstein/": change = 458497]

+++ Code/Game/Gameplay/Actors/Player.cpp
@@ -119,7 +119,7 @@ void Player::SwitchWeapon( int weaponToSwitchTo )
 	Actor* possessedActor = GetActor();
 	Weapons actorWeapons = possessedActor->m_weapons;
 
-	if( weaponToSwitchTo >= actorWeapons.size() )
+	if( weaponToSwitchTo >= static_cast<int>( actorWeapons.size() ) )
 	{
 		weaponToSwitchTo = 0;
 	}
